Tried optimizing size : reduced videobuffers and changed types

+++ sw/applications/0_DOOM/doom/hu_stuff.c
@@ -106,7 +106,7 @@ static boolean          message_nottobefuckedwith;
 
 static hu_stext_t       w_message;
 static hu_stext_t       w_fps;
-static int              message_counter;
+static uint8_t          message_counter;
 
 extern int              showMessages;
 static boolean          headsupactive = false;
